pyproject.toml
[project]
name = "agenticarchitecture"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "beautifulsoup4>=4.14.3",
    "chromium>=0.0.0",
    "docx>=0.2.4",
    "ipython>=9.11.0",
    "langchain>=1.2.12",
    "langchain-community>=0.4.1",
    "langchain-core>=1.2.19",
    "langchain-ollama>=1.0.1",
    "langchain-openai>=1.1.11",
    "langchain-tavily>=0.2.17",
    "langchain-text-splitters>=1.1.1",
    "langgraph>=1.1.2",
    "pdfplumber>=0.11.9",
    "playwright>=1.58.0",
    "pyppeteer>=0.0.25",
    "python-docx>=1.2.0",
    "python-dotenv>=1.2.2",
    "requests>=2.32.5",
    "tavily-python>=0.7.23",
    "unstructured>=0.21.5",
    "websockets==10.4",
]
